+++ bedrock/Cargo.toml
@@ -37,3 +37,7 @@ alloy = { version = "1.0.6", default-features = false, features = ["json", "cont
 dotenvy = "0.15.7"
 tokio = "1.45.0"
 tokio-test = "0.4.4"
+
+[features]
+default = ["tooling_tests"]
+tooling_tests = []
